% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crcmodel.R
\name{crcmodel}
\alias{crcmodel}
\title{R6 Class Representing a `crcmodel`}
\description{
This class implements a `crcmodel`.
}
\seealso{
set_natural_history_fn

set_natural_history_fn

set_screening_fn
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{is a character string representing the model name (e.g. CRC-SPIN v 2.1)}

\item{\code{inputs}}{is a list of model inputs. One can add inputs to the object with the set_input function.}

\item{\code{inputs_table}}{is a data.frame listing all inputs added to the model.}

\item{\code{posterior_params}}{is a data.frame containing parameters from the posterior distribution of the model. This table is set with the set_posterior function.}

\item{\code{natural_history_results}}{is the object that will receive results from the natural history function.}

\item{\code{screening_results}}{is the object that will receive results from the natural history function.}

\item{\code{simulate_natural_history_fn}}{is a function that will take as parameters self and ..., simulate the natural history of crc and returns invisible(self).}

\item{\code{simulate_screening_fn}}{is a function that will take as parameters self and ..., simulate the natural history of crc and returns invisible(self).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{crcmodel$new()}}
\item \href{#method-set_input}{\code{crcmodel$set_input()}}
\item \href{#method-to_json}{\code{crcmodel$to_json()}}
\item \href{#method-set_inputs_from_json}{\code{crcmodel$set_inputs_from_json()}}
\item \href{#method-set_posterior}{\code{crcmodel$set_posterior()}}
\item \href{#method-set_natural_history_fn}{\code{crcmodel$set_natural_history_fn()}}
\item \href{#method-set_screening_fn}{\code{crcmodel$set_screening_fn()}}
\item \href{#method-simulate_natural_history}{\code{crcmodel$simulate_natural_history()}}
\item \href{#method-simulate_screening}{\code{crcmodel$simulate_screening()}}
\item \href{#method-clone}{\code{crcmodel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new `crcmodel` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crcmodel$new(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of the model to be created.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
s new `crcmodel` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_input"></a>}}
\if{latex}{\out{\hypertarget{method-set_input}{}}}
\subsection{Method \code{set_input()}}{
Set Input
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crcmodel$set_input(name, value, type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string defining the input name}

\item{\code{value}}{input value. Can be a single value, a list or a vector.}

\item{\code{type}}{Functionality TBD, but inputs can be filtered by type.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Use this function to add a new input to a crcmodel object.
All model inputs shoulhd be added or modified through this function. Inputs can be vectors or lists of strings, numeric or integers, to guarantee that they can be translated to and from JSON without any issues.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-to_json}{}}}
\subsection{Method \code{to_json()}}{
Converts a `crcmodel` to a JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crcmodel$to_json(input_types)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_types}}{vector of input types to include in the json object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a JSON string containing the crcmodel objects that should be exported
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_inputs_from_json"></a>}}
\if{latex}{\out{\hypertarget{method-set_inputs_from_json}{}}}
\subsection{Method \code{set_inputs_from_json()}}{
Set model Inputs from JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crcmodel$set_inputs_from_json(json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{json}}{a JSON string generated by the model_to_json function}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Use this function to set model inputs from a JSON string. Note that the posterior distribution is not included in the json strong of model inputs.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_posterior"></a>}}
\if{latex}{\out{\hypertarget{method-set_posterior}{}}}
\subsection{Method \code{set_posterior()}}{
Set Posterior distribution of model parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crcmodel$set_posterior(
  posteriors_list,
  posterior_weights,
  cols_to_ignore = NULL,
  n_posterior = 1000,
  use_average = F,
  seed = 12345678,
  resample = T
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{posteriors_list}}{named list of one more more data.frames containing the posterior of model parameters.}

\item{\code{posterior_weights}}{character indicating the name of the column that contain weights to be used when sampling from the posterior}

\item{\code{cols_to_ignore}}{character vector of columns name to ignore. This is useful when the posterior files contain columns that are not parameters and are not inputs to the model.}

\item{\code{n_posterior}}{the size of the sample to take from each posterior file.}

\item{\code{use_average}}{T if one wants to use the average value of all parameters rather than the mean}

\item{\code{seed}}{random seed to use when sampling from the posterior}

\item{\code{resample}}{if T, samples from the posterior using the posterior weights. Otherwise, returns the full posterior, and preservers the weights.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Use this function to add a new input to a crcmodel object.
All model inputs shoulhd be added or modified through this function. Inputs can be vectors or lists of strings, numeric or integers, to guarantee that they can be translated to and from JSON without any issues.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_natural_history_fn"></a>}}
\if{latex}{\out{\hypertarget{method-set_natural_history_fn}{}}}
\subsection{Method \code{set_natural_history_fn()}}{
Set Natural History Function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crcmodel$set_natural_history_fn(
  natural_history_fn,
  returns_standardized_df = T
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{natural_history_fn}}{an R function that is capable of simulating the model's natural history.}

\item{\code{returns_standardized_df}}{if TRUE, your natural history function must return a standardized dataframe.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This function allows the user to have a custom natural history function. Having this flexibility effectively decouples the model itself from the package that is used to simulate the model. As in a regression, the functional form of the model is not baked into `lm`. Instead, `lm` allows you to create *any* linear model.
Use this function to define a natural history function. You are advised to create a simple natural history function and put a `browser()` inside it. Then, after creating your model with `my_model <- crcmodel$new()` and adding inputs with `my_model$set_input()`, assign your custom natural history function with `my_model$set_natural_history(my_function_name)`. `my_function_name` should the name of a function that you already sourced from a script or from a package. Finally, run `my_model$simulate_natural_history()` to run your natural history model. You will then see all your inputs inside the `self` object, which is an environment that contains all the inputs for your model.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_screening_fn"></a>}}
\if{latex}{\out{\hypertarget{method-set_screening_fn}{}}}
\subsection{Method \code{set_screening_fn()}}{
Set Screening Function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crcmodel$set_screening_fn(screening_fn, returns_standardized_df = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{screening_fn}}{an R function that is capable of simulating the model's screening.}

\item{\code{returns_standardized_df}}{if TRUE, your natural history function must return a standardized dataframe.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This function allows the user to have a custom screening function.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulate_natural_history"></a>}}
\if{latex}{\out{\hypertarget{method-simulate_natural_history}{}}}
\subsection{Method \code{simulate_natural_history()}}{
Simulate Natural History Wrapper Function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crcmodel$simulate_natural_history(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{any set of parameters passed to this function will be passed along to the user natural history function.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This function is a wrapper around the user natural history function. It passes the `self` object and any other parameters provided to the function.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulate_screening"></a>}}
\if{latex}{\out{\hypertarget{method-simulate_screening}{}}}
\subsection{Method \code{simulate_screening()}}{
Simulate Screening Wrapper Function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crcmodel$simulate_screening(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{any set of parameters passed to this function will be passed along to the user screening function.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This function is a wrapper around the user screening. It passes the `self` object and any other parameters provided to the function.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crcmodel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
