% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crcrdm.R
\docType{package}
\name{crcrdm}
\alias{crcrdm}
\title{crcrdm: Tools for Robust Decision Making analyses for CRC models}
\description{
The crcrdm package contains two main types of objects: crcmodel and crcexperiment
}
\section{crcmodel functions}{

A crcmodel ...
}

\section{crcexperiment functions}{

A crcexperiment ...
}

\section{EMEWS Integration}{

Once a crcexperiment is defined, there are a few options to run the experiment.
}

