% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crcexperiment_blocks.R
\name{get_people_in_block}
\alias{get_people_in_block}
\title{Find IDS of persons in a given block}
\usage{
get_people_in_block(person_ids, blocks, block_id)
}
\arguments{
\item{person_ids}{a vector of integers of people ids. This vector starts in 1.}

\item{blocks}{the number of population blocks.}

\item{block_id}{the id of the population block to query}
}
\value{
a vector of people who belong to the block_id
}
\description{
Find IDS of persons in a given block
}
