% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crcexperiment.R
\name{is.crcexperiment}
\alias{is.crcexperiment}
\title{Checks if object is a `crcexperiment`.}
\usage{
is.crcexperiment(x)
}
\arguments{
\item{x}{the object}
}
\value{
TRUE if object is a `crcexperiment`
}
\description{
Checks if object is a `crcexperiment`.
}
