% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crcmodel.R
\name{is.crcmodel}
\alias{is.crcmodel}
\title{Checks if object is a `crcmodel`.}
\usage{
is.crcmodel(x)
}
\arguments{
\item{x}{the model object}
}
\value{
TRUE if object is a crcmodel
}
\description{
Checks if object is a `crcmodel`.
}
