% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crcexperiment.R
\name{crcexperiment}
\alias{crcexperiment}
\title{R6 Class Representing a `crcexperiment`}
\description{
This class implements a `crcexperiment`.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{models}}{is a list containing crcmodel objects.}

\item{\code{experimental_design}}{is a table containing one row per experiment to be ran.}

\item{\code{grid}}{is a table containing one row per point in the grid experimental design.}

\item{\code{lhs}}{is a table containing one row per point in the lhs experimental design.}

\item{\code{posteriors}}{is a table containing one row per parameter set defined in the posterior of each model included in the experiment.}

\item{\code{experimental_parameters}}{is a list containing details about each experimental parameter. Experimental parameters can be either policy levers or uncertainties. Defining this distinction is up to the user.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{crcexperiment$new()}}
\item \href{#method-set_parameter}{\code{crcexperiment$set_parameter()}}
\item \href{#method-set_design}{\code{crcexperiment$set_design()}}
\item \href{#method-clone}{\code{crcexperiment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new `crcexperiment` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crcexperiment$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{set of crcmodels to be included in the experiment.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
s new `crcexperiment` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-set_parameter}{}}}
\subsection{Method \code{set_parameter()}}{
Set Parameter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crcexperiment$set_parameter(
  parameter_name,
  experimental_design,
  values,
  min,
  max
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameter_name}}{character string defining the parameter name.}

\item{\code{experimental_design}}{Either "grid" or "lhs" Use lhs if you want to create a Latin Hypercube Sample within the min and max bounds you provided. Use Grid}

\item{\code{values}}{use when experimental_design = "grid". This should be a vector including the values to be included in a grid experimental design. Please use parameters and values that can be converted to strings without any issues.}

\item{\code{min}}{use when experimental_design = "lhs". This should bea numeric value indicating the minimum bound in the Latin Hypercube sample.}

\item{\code{max}}{use when experimental_design = "lhs". This should bea numeric value indicating the minimum bound in the Latin Hypercube sample.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This function constructs the experimental_parameter object, and appends experimental parameters that will be visible inside the model in the future.
Experimental parameters can be either uncertainties or decision levers.
Every parameter defined in this function can be accessed within the model by using \code{experimental_parameters$param_name}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_design"></a>}}
\if{latex}{\out{\hypertarget{method-set_design}{}}}
\subsection{Method \code{set_design()}}{
Set Experimental Design
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crcexperiment$set_design(
  n_lhs,
  convert_lhs_to_grid = F,
  lhs_to_grid_midpoints = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_lhs}}{The number of points in the Latin Hypercube Sample to be created.}

\item{\code{convert_lhs_to_grid}}{Default is FALSE. If TRUE, this function convert the LHS parameters to a grid design. Often is used when testing the "corners" of the experimental design before performing a full LHS run.}

\item{\code{lhs_to_grid_midpoints}}{Only relevant when convert_to_lhs = T. Default value is 0. This should be an integer determining how many points within the grid hypercube should be created for the parameters being converted from LHS to a GRID design. For example, if convert_lhs_to_grid = T and lhs_to_grid_midpoints = 0, this function will create a full factorial design of the LHS parameters with 2^n points. If one wants to use one midpoint, then the design will have 3^n points, and so on. This parameter does not affect parameters orignally defined as part of a grid design because their values have already been set.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Creates the future_experimental_design data.frame based on the paramers defined by the set_parameter functions. The experimental design created by this function is useful to run a typical RDM analysis where each policy is evaluated across a LHS of deep uncertainties. To achieve that, define each policy lever as a grid parameter, and each uncertainty as an "lhs" uncertainty.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crcexperiment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
