% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crcrdm.R
\docType{package}
\name{crcrdm}
\alias{crcrdm}
\title{crcrdm: Tools for Robust Decision Making analyses for CRC models}
\description{
This package implements the \code{\link{crcmodel}} and the \code{\link{crcexperiment}} R6 classes and can be used to perform large-scale computational experiments of multiple cancer screening models using High-Performance Computing resources. The package supports common tasks, such as defining and conducting Probabilistic Sensitivity Analyses (PSAs) and robustness analyses using one or multiple models. The package is also designed to support High-Performance Computing workflows using the Extreme Scale Model Exploration with Swift/T (EMEWS) framework. This feature will allow use of High-Performance computing resources for large-scale, long-running experiments that involve simulating a large populations (e.g. 10^7) over thousands of simulation runs, which require several thousand computing hours and many terabytes of memory to run. This package has been primarly used by the CRC-SPIN model so far, but could be used for other models. The package itself does not contain any specific model.
}
